Conflicts:
	src/com/flabser/server/WebServer.java

+++ rule/CashTracker/Resources/scripts/cashtracker/test/dao/AccountTest.java
@@ -1,5 +1,7 @@
 package cashtracker.test.dao;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
@@ -11,6 +13,8 @@ import org.junit.Test;
 
 import cashtracker.dao.AccountDAO;
 import cashtracker.model.Account;
+import cashtracker.validation.AccountValidator;
+import cashtracker.validation.ValidationError;
 
 import com.flabser.dataengine.pool.DatabasePoolException;
 import com.flabser.util.Util;
@@ -19,6 +23,7 @@ import com.flabser.util.Util;
 public class AccountTest extends InitEnv {
 
 	AccountDAO dao;
+	private AccountValidator validator = new AccountValidator();
 
 	@Override
 	@Before
@@ -30,8 +35,6 @@ public class AccountTest extends InitEnv {
 
 	@Test
 	public void insertTest() {
-		assertNotNull(db);
-
 		int size = dao.findAll().size();
 		int iteration = size + 2;
 
@@ -52,6 +55,13 @@ public class AccountTest extends InitEnv {
 			m.setEnabled(Util.getRandomBoolean());
 			m.setNote("note " + i);
 
+			ValidationError ve = validator.validate(m);
+			if (ve.hasError()) {
+				for (cashtracker.validation.ValidationError.Error err : ve.getErrors()) {
+					assertFalse("ValidationError : " + err.toString(), ve.hasError());
+				}
+			}
+
 			dao.add(m);
 		}
 	}
@@ -69,12 +79,26 @@ public class AccountTest extends InitEnv {
 	}
 
 	@Test
+	public void findAllEnabledTest() {
+		List <Account> list = dao.findAllEnabled();
+		assertTrue(list.size() > 0);
+	}
+
+	@Test
+	public void existsTransactionByAccountTest() {
+		List <Account> accounts = dao.findAll();
+		assertTrue(dao.existsTransactionByAccount(accounts.get(0)));
+	}
+
+	@Test
 	public void updateTest() {
 		List <Account> list = dao.findAll();
 
 		for (Account m : list) {
-			m.setName(m.getName() + "-u");
-			System.out.println(dao.update(m));
+			String name = m.getName() + "-u";
+			m.setName(name);
+			dao.update(m);
+			assertEquals(m.getName(), name);
 		}
 	}
 
@@ -83,7 +107,9 @@ public class AccountTest extends InitEnv {
 		List <Account> list = dao.findAll();
 
 		for (Account m : list) {
-			dao.delete(m);
+			if (!dao.existsTransactionByAccount(m)) {
+				dao.delete(m);
+			}
 		}
 	}
 }
